move Propagation#propagate(...) into AstBasedPropgation
